+++ src/main/kotlin/com/automatemic/backend/Booking/BookingController.kt
@@ -1,5 +1,6 @@
 package com.automatemic.backend.Booking
 
+import com.automatemic.backend.Common.Mappers.BookingDto
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
