+++ src/main/kotlin/com/automatemic/backend/Booking/BookingRepository.kt
@@ -1,5 +1,6 @@
 package com.automatemic.backend.Booking
 
+import com.automatemic.backend.Common.Models.Booking
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
