+++ src/main/kotlin/com/automatemic/backend/Booking/BookingRepositoryCustom.kt
@@ -1,5 +1,7 @@
 package com.automatemic.backend.Booking
 
+import com.automatemic.backend.Common.Models.Booking
+
 interface BookingRepositoryCustom {
     fun findByVenueId(id: Long): List<Booking>
     fun findByCustomerId(id: Long): List<Booking>
