+++ src/main/kotlin/com/automatemic/backend/Booking/BookingRepositoryCustomImpl.kt
@@ -1,5 +1,6 @@
 package com.automatemic.backend.Booking
 
+import com.automatemic.backend.Common.Models.Booking
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Repository
 import javax.persistence.EntityManager
