+++ src/main/kotlin/com/automatemic/backend/Booking/BookingService.kt
@@ -1,5 +1,7 @@
 package com.automatemic.backend.Booking
 
+import com.automatemic.backend.Common.Mappers.BookingDto
+import com.automatemic.backend.Common.Models.Booking
 import com.automatemic.backend.Customer.CustomerRepository
 import com.automatemic.backend.Venue.VenueRepository
 import org.springframework.beans.factory.annotation.Autowired
