+++ src/main/kotlin/com/automatemic/backend/Common/Mappers/CustomerDto.kt
@@ -1,5 +1,6 @@
-package com.automatemic.backend.Customer
+package com.automatemic.backend.Common.Mappers
 
+import com.automatemic.backend.Common.Models.Customer
 import lombok.AllArgsConstructor
 import lombok.Data
 import lombok.NoArgsConstructor
