+++ src/main/kotlin/com/automatemic/backend/Customer/CustomerController.kt
@@ -1,5 +1,6 @@
 package com.automatemic.backend.Customer
 
+import com.automatemic.backend.Common.Mappers.CustomerDto
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
