+++ src/main/kotlin/com/automatemic/backend/Customer/CustomerRepository.kt
@@ -1,5 +1,6 @@
 package com.automatemic.backend.Customer
 
+import com.automatemic.backend.Common.Models.Customer
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
