+++ src/main/kotlin/com/automatemic/backend/Customer/CustomerService.kt
@@ -1,5 +1,7 @@
 package com.automatemic.backend.Customer
 
+import com.automatemic.backend.Common.Mappers.CustomerDto
+import com.automatemic.backend.Common.Mappers.toCustomerEntity
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
