+++ src/main/kotlin/com/automatemic/backend/Venue/VenueController.kt
@@ -1,5 +1,6 @@
 package com.automatemic.backend.Venue
 
+import com.automatemic.backend.Common.Mappers.VenueDto
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.PathVariable
