+++ src/main/kotlin/com/automatemic/backend/Venue/VenueRepository.kt
@@ -1,5 +1,6 @@
 package com.automatemic.backend.Venue
 
+import com.automatemic.backend.Common.Models.Venue
 import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.stereotype.Repository
