+++ src/main/kotlin/com/automatemic/backend/Venue/VenueService.kt
@@ -1,5 +1,7 @@
 package com.automatemic.backend.Venue
 
+import com.automatemic.backend.Common.Mappers.VenueDto
+import com.automatemic.backend.Common.Mappers.toVenueEntity
 import org.springframework.beans.factory.annotation.Autowired
 import org.springframework.stereotype.Service
