+++ src/main/resources/application.properties
@@ -5,8 +5,8 @@ spring.jpa.hibernate.ddl-auto=none
 
 spring.datasource.initialization-mode=always
 spring.datasource.platform=postgres
-spring.datasource.url=jdbc:postgresql://localhost:5432/automateMic
+spring.datasource.url=jdbc:postgresql://localhost:5432/automatemic
 spring.datasource.username=admin
-spring.datasource.password=password
-
+spring.datasource.password=admin
+server.port = 4000
 spring.jpa.properties.hibernate.jdbc.lob.non_contextual_creation=true
